+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 
 - Foundation constants: `NSNotFound`.
 
+- Foundation error codes.
+
 - `core_graphics` module for [Core Graphics](https://developer.apple.com/documentation/coregraphics)
   framework.